itemデータクラスをdata/model配下に分割
